refactoring job search finder
