+++ styles/pages/JobViz/job-viz-page.scss
@@ -1,4 +1,5 @@
 @import './sections/jobCategories.scss';
+@import './modals/selectedJobStyles.scss';
 
 $heroColor: #f0f4ff;
 $jobVizCategoryChainTxt: #6812d1;
