+++ styles/pages/JobViz/modals/selectedJobStyles.scss
@@ -1,0 +1,3 @@
+.selectedJobModal{
+    border: solid pink !important;
+}
